Fix(sync): de-dup sync files before starting
- This is the only way to handle how google deals with conflicts (it basically just creates a bunch of duplicates in files or directories suffixed with a number, e.g " (1)".
- Files are compared using full decrypted content - this gives me the option to do something interesting with newline conversions at some point.  Though,
I probably want to use a hash - that which would mean line endings would need to be canonicalized prior to encryption.
- The dups are eliminated in a consistent order, that is, the file with the lowest (alphabetical) revguid is preserved.  if there is more than one dup with the same revguid,
then filename ordering is used.  This means that a machine might consider a "dup" to be a file it created; it has to be so, otherwise ALL the sync files for a given native file could be removed.
- This change has implications for the rest of the code that I haven't dealt with yet, so things are not working right now.